+++ airflow/utils/sqlalchemy.py
@@ -19,13 +19,9 @@
 import datetime
 import json
 import logging
-import os
-import time
-import traceback
 
 import pendulum
 from dateutil import relativedelta
-from sqlalchemy import event, exc
 from sqlalchemy.types import DateTime, Text, TypeDecorator
 
 from airflow.configuration import conf
@@ -37,60 +33,7 @@ utc = pendulum.tz.timezone('UTC')
 using_mysql = conf.get('core', 'sql_alchemy_conn').lower().startswith('mysql')
 
 
-def setup_event_handlers(engine):
-    """
-    Setups event handlers.
-    """
-    # pylint: disable=unused-argument
-    @event.listens_for(engine, "connect")
-    def connect(dbapi_connection, connection_record):
-        connection_record.info['pid'] = os.getpid()
-
-    if engine.dialect.name == "sqlite":
-        @event.listens_for(engine, "connect")
-        def set_sqlite_pragma(dbapi_connection, connection_record):
-            cursor = dbapi_connection.cursor()
-            cursor.execute("PRAGMA foreign_keys=ON")
-            cursor.close()
-
-    # this ensures sanity in mysql when storing datetimes (not required for postgres)
-    if engine.dialect.name == "mysql":
-        @event.listens_for(engine, "connect")
-        def set_mysql_timezone(dbapi_connection, connection_record):
-            cursor = dbapi_connection.cursor()
-            cursor.execute("SET time_zone = '+00:00'")
-            cursor.close()
-
-    @event.listens_for(engine, "checkout")
-    def checkout(dbapi_connection, connection_record, connection_proxy):
-        pid = os.getpid()
-        if connection_record.info['pid'] != pid:
-            connection_record.connection = connection_proxy.connection = None
-            raise exc.DisconnectionError(
-                "Connection record belongs to pid {}, "
-                "attempting to check out in pid {}".format(connection_record.info['pid'], pid)
-            )
-    if conf.getboolean('debug', 'sqlalchemy_stats', fallback=False):
-        @event.listens_for(engine, "before_cursor_execute")
-        def before_cursor_execute(conn, cursor, statement, parameters, context, executemany):
-            conn.info.setdefault('query_start_time', []).append(time.time())
-
-        @event.listens_for(engine, "after_cursor_execute")
-        def after_cursor_execute(conn, cursor, statement, parameters, context, executemany):
-            total = time.time() - conn.info['query_start_time'].pop()
-            file_name = [
-                f"'{f.name}':{f.filename}:{f.lineno}" for f
-                in traceback.extract_stack() if 'sqlalchemy' not in f.filename][-1]
-            stack = [f for f in traceback.extract_stack() if 'sqlalchemy' not in f.filename]
-            stack_info = ">".join([f"{f.filename.rpartition('/')[-1]}:{f.name}" for f in stack][-3:])
-            conn.info.setdefault('query_start_time', []).append(time.monotonic())
-            log.info("@SQLALCHEMY %s |$ %s |$ %s |$  %s ",
-                     total, file_name, stack_info, statement.replace("\n", " ")
-                     )
-
 # pylint: enable=unused-argument
-
-
 class UtcDateTime(TypeDecorator):
     """
     Almost equivalent to :class:`~sqlalchemy.types.DateTime` with
